# swagger.yaml
openapi: 3.0.0
info:
  title: Sample API
  version: 1.0.0
paths:
...
  /user/create:
    post:
      summary: Người dùng đăng kí
      tags:
        - Users
      description: Người dùng đăng kí
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                password:
                  type: string
                phoneNumber:
                  type: string
                address:
                  type: string
                email:
                  type: string
              required:
      responses:
        '201':
          description: Thành công. Trả về Success.
  /user/login:
    post:
      summary: Người dùng đăng nhập
      tags:
        - Users
      description: Người dùng đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                password:
                  type: string
              required: # Không cần xác định trường required ở đây
      responses:
        '201':
          description: Thành công. Trả về jwt.
  /user/receive-voucher:
    post:
      summary: Người dùng nhận voucher
      tags:
        - Users
      description: Người dùng nhận voucher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                voucherId:
                  type: number
              required: # Không cần xác định trường required ở đây
      responses:
        '201':
          description: Thành công. Trả về Success.
  /user/verify-email:
    get:
      summary: Người dùng verify email
      tags:
        - Users
      description: sau khi đăng kí tài khoản. Người dùng cần verify email
      parameters:
        - in: query
          name: token
          schema:
            type: string
          required: false
      responses:
        '201':
          description: Thành công. Trả về Success.
...